requirements.txt
python-telegram-bot==13.11
beautifulsoup4==4.11.1
OpenCC==1.1.3
